auth-service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aerich==0.7.2
    # via -r requirements.in
aiosqlite==0.17.0
    # via tortoise-orm
anyio==3.7.1
    # via
    #   httpcore
    #   starlette
    #   watchfiles
asgi-lifespan==2.1.0
    # via -r requirements.in
async-timeout==4.0.2
    # via redis
asyncpg==0.28.0
    # via tortoise-orm
bcrypt==4.0.1
    # via passlib
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
cffi==1.15.1
    # via cryptography
click==8.1.6
    # via
    #   aerich
    #   redis-om
    #   uvicorn
cryptography==41.0.3
    # via
    #   types-pyopenssl
    #   types-redis
dictdiffer==0.9.0
    # via aerich
dnspython==2.4.1
    # via email-validator
email-validator==2.0.0.post2
    # via fastapi
exceptiongroup==1.1.2
    # via
    #   anyio
    #   pytest
fastapi[all]==0.99.1
    # via
    #   -r requirements.in
    #   fastapi-utils
fastapi-utils==0.2.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hiredis==2.2.3
    # via redis-om
httpcore==0.17.3
    # via httpx
httptools==0.6.0
    # via uvicorn
httpx==0.24.1
    # via fastapi
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.0.0
    # via pytest
iso8601==1.1.0
    # via tortoise-orm
itsdangerous==2.1.2
    # via fastapi
jinja2==3.1.2
    # via fastapi
markupsafe==2.1.3
    # via jinja2
more-itertools==9.1.0
    # via redis-om
orjson==3.9.3
    # via fastapi
packaging==23.1
    # via pytest
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pluggy==1.2.0
    # via pytest
pycparser==2.21
    # via cffi
pydantic==1.10.12
    # via
    #   -r requirements.in
    #   aerich
    #   fastapi
    #   fastapi-utils
    #   redis-om
pyjwt==2.8.0
    # via -r requirements.in
pypika-tortoise==0.1.6
    # via tortoise-orm
pytest==7.4.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.21.1
    # via -r requirements.in
python-dotenv==1.0.0
    # via uvicorn
python-http-client==3.3.7
    # via sendgrid
python-multipart==0.0.6
    # via fastapi
python-ulid==1.1.0
    # via redis-om
pytz==2023.3
    # via tortoise-orm
pyyaml==6.0.1
    # via
    #   fastapi
    #   uvicorn
redis==4.6.0
    # via redis-om
redis-om==0.2.1
    # via -r requirements.in
scheduler==0.8.4
    # via -r requirements.in
sendgrid==6.10.0
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   asgi-lifespan
    #   httpcore
    #   httpx
sqlalchemy==1.4.49
    # via fastapi-utils
starkbank-ecdsa==2.2.0
    # via sendgrid
starlette==0.27.0
    # via fastapi
tomli==2.0.1
    # via pytest
tomlkit==0.12.1
    # via aerich
tortoise-orm[asyncpg]==0.19.3
    # via
    #   -r requirements.in
    #   aerich
typeguard==4.1.0
    # via scheduler
types-pyopenssl==23.2.0.2
    # via types-redis
types-redis==4.6.0.3
    # via redis-om
typing-extensions==4.7.1
    # via
    #   aiosqlite
    #   fastapi
    #   pydantic
    #   redis-om
    #   typeguard
    #   uvicorn
ujson==5.8.0
    # via fastapi
uvicorn[standard]==0.23.2
    # via fastapi
uvloop==0.17.0
    # via uvicorn
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn
